Fix old header rendering
Style tag was being escaped, leaving the CSS
to be output as a string.
This fix allows the style tag to be output
without being stripped by KSES only whilst
the header block is being rendered.
Since the CSS being rendered is properly
escaped in the renderer, this should be
acceptable.