Move petition title truncation to CSS
WP's word-count truncation truncates characters in Thai,
rather than words. Other truncation methods using PHP
string manipulation may work, but it is unnecessary
computation if we can do the same in CSS.
Although the CSS method does use '-webkit-' browser
vendor prefixes, they are supported in all major
browsers (as the spec is still in draft). so this
should allow for a similar experience in modern
browsers, without causing too much layout shift in
browsers that don't support it. So it's a reasonable
compromise in this specific scenario.